requirements.txt
Pillow>=9.0.0
tkinterdnd2>=0.3.0
rarfile>=4.0
pyinstaller>=6.0.0 